aura.cfg: Add the correct 1.27 version exe & hash for eurobattle.net example server configuration.
A big thanks to Phong for sniffing it for me. On how to do this check this thread

  http://www.codelain.com/forum/index.php?topic=441.0

on the original GHost++ forum where an user explains how to get it via Wireshark.
For War3 version 1.27 the bytes are 10 00 1b 01 6e f8 14 57, of which the first
4 bytes go to version exe and the other 4 bytes to version exe hash. Then the final
step is that they are converted to decimal as aura.cfg expects it in decimal format.